fix: add interface to gitops package
