+++ litmus-portal/graphql-server/pkg/database/mongodb/gitops/operations.go
@@ -9,8 +9,20 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
+// Operator is the struct for gitOps operator
+type Operator struct {
+	operator mongodb.MongoOperator
+}
+
+// NewGitOpsOperator returns a new instance of gitOps operator
+func NewGitOpsOperator(mongodbOperator mongodb.MongoOperator) *Operator {
+	return &Operator{
+		operator: mongodbOperator,
+	}
+}
+
 // AddGitConfig inserts new git config for project
-func AddGitConfig(ctx context.Context, config *GitConfigDB) error {
+func (g *Operator) AddGitConfig(ctx context.Context, config *GitConfigDB) error {
 	err := mongodb.Operator.Create(ctx, mongodb.GitOpsCollection, config)
 	if err != nil {
 		return err
@@ -19,7 +31,7 @@ func AddGitConfig(ctx context.Context, config *GitConfigDB) error {
 }
 
 // GetGitConfig retrieves git config using project id
-func GetGitConfig(ctx context.Context, projectID string) (*GitConfigDB, error) {
+func (g *Operator) GetGitConfig(ctx context.Context, projectID string) (*GitConfigDB, error) {
 	query := bson.D{{"project_id", projectID}}
 	var res GitConfigDB
 	result, err := mongodb.Operator.Get(ctx, mongodb.GitOpsCollection, query)
@@ -35,7 +47,7 @@ func GetGitConfig(ctx context.Context, projectID string) (*GitConfigDB, error) {
 }
 
 // GetAllGitConfig retrieves all git configs from db
-func GetAllGitConfig(ctx context.Context) ([]GitConfigDB, error) {
+func (g *Operator) GetAllGitConfig(ctx context.Context) ([]GitConfigDB, error) {
 	query := bson.D{{}}
 	results, err := mongodb.Operator.List(ctx, mongodb.GitOpsCollection, query)
 	if err != nil {
@@ -50,7 +62,7 @@ func GetAllGitConfig(ctx context.Context) ([]GitConfigDB, error) {
 }
 
 // ReplaceGitConfig updates git config matching the query
-func ReplaceGitConfig(ctx context.Context, query bson.D, update *GitConfigDB) error {
+func (g *Operator) ReplaceGitConfig(ctx context.Context, query bson.D, update *GitConfigDB) error {
 	updateResult, err := mongodb.Operator.Replace(ctx, mongodb.GitOpsCollection, query, update)
 	if err != nil {
 		return err
@@ -64,7 +76,7 @@ func ReplaceGitConfig(ctx context.Context, query bson.D, update *GitConfigDB) er
 }
 
 // UpdateGitConfig update git config matching the query
-func UpdateGitConfig(ctx context.Context, query bson.D, update bson.D) error {
+func (g *Operator) UpdateGitConfig(ctx context.Context, query bson.D, update bson.D) error {
 	updateResult, err := mongodb.Operator.Update(ctx, mongodb.GitOpsCollection, query, update)
 	if err != nil {
 		return err
@@ -78,7 +90,7 @@ func UpdateGitConfig(ctx context.Context, query bson.D, update bson.D) error {
 }
 
 // DeleteGitConfig removes git config corresponding to the given project id
-func DeleteGitConfig(ctx context.Context, projectID string) error {
+func (g *Operator) DeleteGitConfig(ctx context.Context, projectID string) error {
 	query := bson.D{{"project_id", projectID}}
 	_, err := mongodb.Operator.Delete(ctx, mongodb.GitOpsCollection, query)
